\name{array}
\alias{array}
\encoding{latin1}
\docType{data}
\title{Combination of multiple array experiments}
\description{Combination of 5 experiments. The data has been YuGene transformed, mapped to Ensembl ID. 2000 genes have been randomly selected.
}
\usage{data(array)}
\format{

A list containing the following components:
  \describe{
    \item{\code{data.all}}{ Matrix of 82 samples and 2000 gene expression.}
    \item{\code{experiment.all}}{a factor containing the name of the experiments.}
    \item{\code{platform.all}}{a factor containing the platform of each sample.}
    \item{\code{type.all}}{a factor containing the type of each sample.}
    }
}

\source{
The data were downloaded from \url{www.stemformatics.org}.
}
\references{
Brennand KJ, Simone A, Jou J, Gelboin-Burkhart C et al. Modelling schizophrenia using human induced pluripotent stem cells. Nature 2011 May 12;473(7346):221-5. PMID: 21490598\cr

Zaehres H, Kögler G, Arauzo-Bravo MJ, Bleidissel M et al. Induction of pluripotency in human cord blood unrestricted somatic stem cells. Exp Hematol 2010 Sep;38(9):809-18, 818.e1-2. PMID: 20541586 \cr

Jia F, Wilson KD, Sun N, Gupta DM et al. A nonviral minicircle vector for deriving human iPS cells. Nat Methods 2010 Mar;7(3):197-9. PMID: 20139967\cr

Maherali N, Ahfeldt T, Rigamonti A, Utikal J et al. A high-efficiency system for the generation and study of human induced pluripotent stem cells. Cell Stem Cell 2008 Sep 11;3(3):340-5. PMID: 18786420 \cr

Nayler S, Gatei M, Kozlov S, Gatti R et al. Induced pluripotent stem cells from ataxia-telangiectasia recapitulate the cellular phenotype. Stem Cells Transl Med 2012 Jul;1(7):523-35. PMID: 23197857\cr

}
\keyword{datasets}
